converse a toggle-buuton in a component
